try sr04 control

+++ Core/Inc/ch34g.h
@@ -1,7 +1,7 @@
 #include <stdint.h>
 
 /* 旋转舵机角度*/
-void CH34G_TurnAngle(uint32_t angle, uint16_t delay);
+void CH34G_TurnAngle(uint32_t angle, uint32_t delay);
 
 /* 初始化舵机 */
 void CH34G_Init();
